Updated logic. Finished add image and registration. Created gif screen popup. Added pods for gif loading.

+++ ImageGallery/API/Models/ImageForm.swift
@@ -12,18 +12,18 @@ protocol ImageFormType {
     var image: Data { get }
     var description: String? { get }
     var hashtag: String? { get }
-    var latitude: Float { get }
-    var longitude: Float { get }
+    var latitude: Double { get }
+    var longitude: Double { get }
 }
 
 class ImageForm: ImageFormType, Decodable {
     var image: Data
     var description: String?
     var hashtag: String?
-    var latitude: Float
-    var longitude: Float
+    var latitude: Double
+    var longitude: Double
     
-    init(image: Data, description: String?, hashtag: String?, latitude: Float, longitude: Float) {
+    init(image: Data, description: String?, hashtag: String?, latitude: Double, longitude: Double) {
         self.image = image
         self.description = description
         self.hashtag = hashtag
